Rename INT_TEST_4HZ and _8HZ to _SLOW and _FAST
Chris Rorden requested a naming improvement.

We miscalculated the speeds initially. According to the
data sheet, the speeds should be a function of the f_CLK:

   f_CLK/2^21 or f_CLK/2^20

Since f_CLK is 2.048 MHz, the names should have been _1HZ
and _2HZ respectively. As we have not verified the actual
speeds, we have decided to name them _SLOW and _FAST.

See also, the ADS1298 and ADS1299 data sheets,
 * REGISTER MAP
   - CONFIG2
 * THEORY OF OPERATION
   - PGA SETTINGS AND INPUT RANGE
     - ADC \Delta\Sigma Modulator