Added Arduino IDE compatible firmware for rev1
The new eegmousefirmware directory contains a modified version of the code
in arduino-src that can be used with the Arduino IDE to load the firmware
onto a rev1 board.

The Arduino IDE places several restrictions on file names: filenames can not
contain dashes, the main file of a project must be named .ino (rather than
.cpp), and the directory name must match the name of the main file.

Several changes were also made to stream data via the SAM3X's native USB port
rather than the Arduino Due's programming port.  This included switching to
SerialUSB and using "setup" and "loop" functions rather than "main".

Finally, this checkin turns on all eight channels by default.  This seemed like
a friendlier default configuration because it is easier for a user to figure
out how to discard data from the channels they don't care about than to figure
out how to modify the firmware to turn on additional channels.