working commandline build for Arduino DUE
When using the "native" (not "programming") port, the process is
not as reliable as I would like. When it fails, I find that I
must go through the steps of hard-resetting the board and waiting
until the serial port can be actually used.

Once a flash is successful, and the board is reset, the firmware
works as expected.