sketch out a start for dealing with different hardware versions
Christoph Maier pointed out that PWDN and RESET were backwards,
and indeed this highlights that rev0 and rev1 versions of our
hardware have different pins for some purposes.

More than just defines, currently have code which initializes
the PINs 50, 15, and 52 which is important for rev0 and harmless
for rev1. It would be better if we did not do confusing things.