add delay for ads1299 power up
It may take a bit longer for the power rails to stabilize on the isolated side of the
board than it takes on the non-isolated side.  This checkin adds a 300 ms delay before
the SAM3X starts communicating with the ADS1299 to make sure the power has had time
to stabilize.