Switch to 2D primitives which are extruded to 3D
Our laser cut parts are 2D patterns that are used to cut a 3D sheet, and
we need these 2D patterns (e.g. DXF files) for the laser cutter.  We
could get these 2D patterns either by taking the projection of a 3D
object, or by creating a 2D object directly.  Either one should work,
but starting with the 2D pattern should be more computationally
efficient and OpenSCAD (version 2013.05.28) is crashing when I try to
take the 2D projection of our current top plate.  This check in thus
replaces the cubes and cylinders with squares and circles to create a 2D
part, and then extrudes these to create the same 3D result as we had
earlier.