Expand screw holes to be proper size for #6 screws
Our design had 3.6 mm screw holes that were a bit small for #6 screws,
but Gold Phoenix PCB appears to have drilled larger 3.8 mm holes for our
prototype boards (which is the size Machinery's Handbook recommends for
a free clearance fit for this screw size).  To save us from problems
with future board runs, this check in expands the holes in the design to
match the 3.8 mm size of the holes on the prototype boards.