Split the case OpenSCAD file into multiple files.
It's useful to be able to render the case as both a 3D model and a 2D
pattern that can be used with a laser cutter.  It seems to be easiest to
support this in OpenSCAD by using one file to render the 3D model
(rev1_case_v1.scad) and a second file to render the 2D pattern
(rev1_case_v1_pattern.scad).  The code describing the case was left in
the rev1_case_v1.scad file, as including the 3D model with the
code describing it seemed most natural.

It also seemed reasonable to split the code to render the stack of
boards for rev1 into a separate file (rev1.scad), as it made the
description of the case design itself much less cluttered and we
might want to use the rev1 board model with other case designs.

Finally, the constants defining various dimensions in the design were
split into separate files (rev1_dimensions.scad and
rev1_case_v1_dimensions.scad).  This was done to work around a
constraint of OpenSCAD, namely that "use" will not import variables, but
"include" will render all of the statements in a file.