Allow kerf to be specified.
The slot cut by a tool like a laser cutter has a finite width (known as
the kerf of the tool), so that holes and slots will be wider than they
are in the pattern and cut out objects will be smaller than they are in
the pattern.  We correct for this by moving the center of the cutting
line inward (for holes) or outward (for cut out shapes) so that the edge
of the cut line will be in the correct place.