Fix draw order.
OpenSCAD does not sort transparent objects from back to front when
drawing, so sometimes you can't see one transparent object behind
another.  The bottom of the case was being draw after the top, so
it wasn't visible through the top of the case when looking from above
(the common case); drawing the bottom first fixes this (but causes
drawing problems when viewing the device from the bottom, the less
common case).